\name{ROCnReg-package}
\alias{ROCnReg-package}
\alias{ROCnReg}
\docType{package}
\title{
\packageTitle{ROCnReg}
}
\description{
\packageDescription{ROCnReg}
}
\details{
\tabular{ll}{
Package: \tab ROCnReg\cr
Type: \tab Package\cr
Version: \tab 1.0-4\cr
Date: \tab 2020-12-06\cr
License: \tab GPL\cr
}
}
\author{
Maria Xose Rodriguez-Alvarez and Vanda Inacio 
Maintainer: Maria Xose Rodriguez-Alvarez <mxrodriguez@bcamath.org>
}
\references{
Erkanli, A., Sung M., Jane Costello, E., and Angold, A. (2006). Bayesian semi-parametric ROC analysis. Statistics in Medicine, \bold{25}, 3905--3928.

Faraggi, D. (2003). Adjusting receiver operating characteristic curves and related indices for covariates. \emph{The Statistician} \bold{52}, 179--192.

Gu, J., Ghosal, S., and Roy, A. (2008). Bayesian bootstrap estimation of ROC curve. Statistics in Medicine, \bold{27}, 5407--5420.

Inacio de Carvalho, V., Jara, A., Hanson, T. E., and de Carvalho, M. (2013). Bayesian nonparametric ROC regression modeling. Bayesian Analysis, \bold{8}, 623--646.

Inacio de Carvalho, V., and Rodriguez-Alvarez, M. X. (2018). Bayesian nonparametric inference for the covariate-adjusted ROC curve. arXiv preprint arXiv:1806.00473.

Janes, H., and Pepe, M.S. (2009). Adjusting for covariate effects on classification accuracy using the covariate-adjusted receiver operating characteristic curve. Biometrika, \bold{96}, 371--382.

Pepe, M.S. (1998). Three approaches to regression analysis of receiver operating characteristic curves for continuous test results. \emph{Biometrics} \bold{54}, 124--135.

Rodriguez-Alvarez, M. X. and Inacio, V., and  (2020). ROCnReg: An R Package for Receiver Operating Characteristic Curve Inference with and without Covariate Information. arXiv preprint arXiv:2003.13111.

Rodriguez-Alvarez, M.X., Tahoces, P. G., Cadarso-Suarez, C., and Lado, M.J. (2011). Comparative study of ROC regression techniques--Applications for the computer-aided diagnostic system in breast cancer detection. Computational Statistics and Data Analysis, \bold{55}, 888--902.

Rodriguez-Alvarez, M.X., Roca-Pardinas, J., and Cadarso-Suarez, C. (2011). ROC curve and covariates: extending induced methodology to the non-parametric framework. Statistics and Computing, \bold{21}, 483--499.
}
%~~ Optionally other standard keywords, one per line, from file KEYWORDS in ~~
%~~ the R documentation directory ~~
\keyword{ package }
%\seealso{
%}
%\examples{
%}
